+++ lib/ansi.mli
@@ -11,7 +11,22 @@ val process : t -> string -> string
     next call to [process]. *)
 
 val css : string
-(** Some default CSS rules to make the HTML output appear in colour. *)
+(** Some default CSS rules to make the HTML output appear in colour, using xterm defaults (light
+    variant). *)
+
+val css_dark : string
+(** Some default CSS rules to make the HTML output appear in colour, using xterm defaults (dark
+    variant). *)
+
+val css_solarized : string
+(** Some default CSS rules to make the HTML output appear in colour, with the Solarized colorscheme
+    (light variant).
+    @see https://ethanschoonover.com/solarized/ *)
+
+val css_solarized_dark : string
+(** Some default CSS rules to make the HTML output appear in colour, with the Solarized colorscheme
+    (dark variant).
+    @see https://ethanschoonover.com/solarized/ *)
 
 val strip : string -> string
 (** [strip data] reads in [data] and strips out ANSI sequences. It doesn't remember partial